[BUGFIX] A bug where the method node lookup didn't work
The 'getMethod' method in the Invoker wrapper did only check the
descriptor of the parameters. To ensure that the right method is
returned, the full method desc must be used.

The return desc (the 'desc' element in the hooks file) is still optional,
but should be used to avoid conflicts.